Database v1.0
At this point all functionality other than secure logins
should be contained within this database with reasonable
keys and data sizes. To use this database, you will need
MySQL 5.1.x and can be imported by MySQL Administrator
into a Schema called INB201. The tables are empty and
are merely there for the form. Actual Peer Review forms
are intended to be stored in an xml Document, although
using this method may break the OS independancy of the
system as Mac/Linux use different directory structures
to Windows. So this may be an issue.

Signed-off-by: Edward Evans <edjojob@gmail.com>